requirements.txt
Flask==2.2.5
Pillow==11.1.0
